+++ lib/rabl-rails.rb
@@ -43,6 +43,12 @@ module RablRails
   mattr_accessor :enable_jsonp_callbacks
   @@enable_jsonp_callbacks = false
 
+  mattr_accessor :allow_empty_format_in_template
+  @@allow_empty_format_in_template = false
+
+  mattr_accessor :xml_options
+  @@xml_options = { dasherize: true, skip_types: false }
+
   def self.configure
     yield self